+++ tests/utils.ts
@@ -24,17 +24,48 @@ function getRandomId() {
   return String(Math.floor(Math.random() * (max - min + 1)) + min);
 }
 
+export const makeRequest = async (app: Express, path: string) => {
+  // @ts-ignore
+  const res = await request(app).get(path);
+  return res;
+};
+
 export const sendTestRequests = async (app: Express, num: number) => {
   for (let index = 0; index < num; index++) {
     const id = getRandomId();
     try {
-      // @ts-ignore
-      const res = await request(app).get(`/api/${id}`);
+      await makeRequest(app, `/api/${id}`);
     } catch (err) {
       throw new Error(err);
     }
   }
   const id = getRandomId();
-  // @ts-ignore
-  await request(app).get(`/api/router/${id}`);
+  try {
+    await makeRequest(app, `/api/router/${id}`);
+  } catch (err) {
+    throw new Error(err);
+  }
+};
+
+export const sendTestRequestNextJS = async (app: Express, num: number) => {
+  const endpoints = [
+    '/',
+    '/users',
+    '/users/:id',
+    '/app-apis',
+    '/app-apis/:id',
+    '/api/hello',
+    '/api/auth/login',
+    '/api/auth/register'
+  ];
+
+  const randomIndex = Math.floor(Math.random() * endpoints.length);
+  let endpoint = endpoints[randomIndex];
+
+  if (endpoint.includes(':id')) {
+    const randomId = Math.floor(Math.random() * 100);
+    endpoint = endpoint.replace(':id', randomId.toString());
+  }
+
+  await makeRequest(app, endpoint);
 };